Removing unnecessary bits and restructuring directory
